app/routes/pet_routes.py
from flask import Blueprint, request, abort, make_response
from ..db import db
from ..models.pet import Pet

import google.generativeai as genai
import os

bp = Blueprint("pets", __name__, url_prefix="/pets")

@bp.post("")
def create_pet():
    pass

@bp.get("")
def get_pets():
    pet_query = db.select(Pet)

    pets = db.session.scalars(pet_query)
    response = []

    for pet in pets:
        response.append(pet.to_dict())

    return response

@bp.get("/<pet_id>")
def get_single_pet(pet_id):
    pet = validate_model(Pet,pet_id)
    return pet.to_dict()

def generate_greetings(character):
    model = genai.GenerativeModel("gemini-1.5-flash")
    input_message = f"I am writing a fantasy RPG video game. I have an npc named {character.name} who is {character.age} years old. They are a {character.occupation} who has a {character.personality} personality. Please generate a Python style list of 10 stock phrases they might use when the main character talks to them. Please return just the list without a variable name and square brackets."
    response = model.generative_content(input_message)
    response_split = response.text.split("\n")
    return response_split[:-1]

def validate_model(cls,id):
    try:
        id = int(id)
    except:
        response =  response = {"message": f"{cls.__name__} {id} invalid"}
        abort(make_response(response , 400))

    query = db.select(cls).where(cls.id == id)
    model = db.session.scalar(query)
    if model:
        return model

    response = {"message": f"{cls.__name__} {id} not found"}
    abort(make_response(response, 404))