requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
cachetools==5.5.0
certifi==2024.2.2
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
Flask==3.0.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.0.3
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
packaging==23.2
pluggy==1.4.0
proto-plus==1.25.0
protobuf==5.29.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.2.0
pytest==8.0.0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.25
tqdm==4.66.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.0.1
